+++ views/cats/_form.php
@@ -35,7 +35,7 @@ use wdmg\widgets\AliasInput;
                 'enctype' => 'multipart/form-data'
             ]
         ]); ?>
-        <?= $form->field($model, 'name')->textInput(['maxlength' => true]) ?>
+        <?= $form->field($model, 'name')->textInput(['maxlength' => true, 'lang' => ($model->locale ?? Yii::$app->language)]) ?>
 
         <?= $form->field($model, 'alias')->widget(AliasInput::class, [
             'labels' => [
@@ -48,7 +48,7 @@ use wdmg\widgets\AliasInput;
         ])->label(Yii::t('app/modules/blog', 'Category URL')); ?>
 
         <?php
-            if (isset(Yii::$app->redirects) && $model->url && ($model->status == $model::STATUS_PUBLISHED)) {
+            if (isset(Yii::$app->redirects) && $model->url) {
                 if ($url = Yii::$app->redirects->check($model->url, false)) {
                     echo Html::tag('div', Yii::t('app/modules/redirects', 'For this URL is active redirect to {url}', [
                         'url' => $url
@@ -69,9 +69,9 @@ use wdmg\widgets\AliasInput;
             </div>
             <div id="postMetaTags" class="panel-collapse collapse">
                 <div class="panel-body">
-                    <?= $form->field($model, 'title')->textInput() ?>
-                    <?= $form->field($model, 'description')->textarea(['rows' => 3]) ?>
-                    <?= $form->field($model, 'keywords')->textarea(['rows' => 3]) ?>
+                    <?= $form->field($model, 'title')->textInput(['lang' => ($model->locale ?? Yii::$app->language)]) ?>
+                    <?= $form->field($model, 'description')->textarea(['rows' => 3, 'lang' => ($model->locale ?? Yii::$app->language)]) ?>
+                    <?= $form->field($model, 'keywords')->textarea(['rows' => 3, 'lang' => ($model->locale ?? Yii::$app->language)]) ?>
                 </div>
             </div>
         </div>