+++ views/cats/index.php
@@ -49,7 +49,7 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
                         ]);
                     }
 
-                    if (isset(Yii::$app->redirects) && $model->url && ($model->status == $model::STATUS_PUBLISHED)) {
+                    if (isset(Yii::$app->redirects) && $model->url) {
                         if ($url = Yii::$app->redirects->check($model->url, false)) {
                             $output .= '&nbsp' . Html::tag('span', '', [
                                 'class' => "text-danger fa fa-exclamation-circle",