+++ views/posts/_form.php
@@ -40,7 +40,7 @@ use wdmg\widgets\AliasInput;
         ]
     ]); ?>
     <div class="col-xs-12 col-sm-12 col-md-8 col-lg-9">
-        <?= $form->field($model, 'name')->textInput(['maxlength' => true]) ?>
+        <?= $form->field($model, 'name')->textInput(['maxlength' => true, 'lang' => ($model->locale ?? Yii::$app->language)]) ?>
 
         <?= $form->field($model, 'alias')->widget(AliasInput::class, [
             'labels' => [
@@ -64,10 +64,11 @@ use wdmg\widgets\AliasInput;
             }
         ?>
 
-        <?= $form->field($model, 'excerpt')->textarea(['rows' => 3]) ?>
+        <?= $form->field($model, 'excerpt')->textarea(['rows' => 3, 'lang' => ($model->locale ?? Yii::$app->language)]) ?>
         <?= $form->field($model, 'content')->widget(Editor::class, [
             'options' => [
                 'id' => 'posts-form-content',
+                'lang' => ($model->locale ?? Yii::$app->language)
             ],
             'pluginOptions' => []
         ]) ?>
@@ -81,9 +82,9 @@ use wdmg\widgets\AliasInput;
             </div>
             <div id="postMetaTags" class="panel-collapse collapse">
                 <div class="panel-body">
-                    <?= $form->field($model, 'title')->textInput() ?>
-                    <?= $form->field($model, 'description')->textarea(['rows' => 3]) ?>
-                    <?= $form->field($model, 'keywords')->textarea(['rows' => 3]) ?>
+                    <?= $form->field($model, 'title')->textInput(['lang' => ($model->locale ?? Yii::$app->language)]) ?>
+                    <?= $form->field($model, 'description')->textarea(['rows' => 3, 'lang' => ($model->locale ?? Yii::$app->language)]) ?>
+                    <?= $form->field($model, 'keywords')->textarea(['rows' => 3, 'lang' => ($model->locale ?? Yii::$app->language)]) ?>
                 </div>
             </div>
         </div>