+++ views/tags/_form.php
@@ -36,7 +36,7 @@ use wdmg\widgets\AliasInput;
                 'enctype' => 'multipart/form-data'
             ]
         ]); ?>
-        <?= $form->field($model, 'name')->textInput(['maxlength' => true]) ?>
+        <?= $form->field($model, 'name')->textInput(['maxlength' => true, 'lang' => ($model->locale ?? Yii::$app->language)]) ?>
 
         <?= $form->field($model, 'alias')->widget(AliasInput::class, [
             'labels' => [
@@ -70,9 +70,9 @@ use wdmg\widgets\AliasInput;
             </div>
             <div id="postMetaTags" class="panel-collapse collapse">
                 <div class="panel-body">
-                    <?= $form->field($model, 'title')->textInput() ?>
-                    <?= $form->field($model, 'description')->textarea(['rows' => 3]) ?>
-                    <?= $form->field($model, 'keywords')->textarea(['rows' => 3]) ?>
+                    <?= $form->field($model, 'title')->textInput(['lang' => ($model->locale ?? Yii::$app->language)]) ?>
+                    <?= $form->field($model, 'description')->textarea(['rows' => 3, 'lang' => ($model->locale ?? Yii::$app->language)]) ?>
+                    <?= $form->field($model, 'keywords')->textarea(['rows' => 3, 'lang' => ($model->locale ?? Yii::$app->language)]) ?>
                 </div>
             </div>
         </div>