+++ views/tags/view.php
@@ -30,6 +30,9 @@ if ($model->locale && isset(Yii::$app->translations) && class_exists('\wdmg\tran
             [
                 'attribute' => 'name',
                 'format' => 'raw',
+                'contentOptions' => [
+                    'lang' => ($model->locale ?? Yii::$app->language)
+                ],
                 'value' => function($model) {
                     $output = Html::tag('strong', $model->name);
                     if (($tagURL = $model->getTagUrl(true, true)) && $model->id) {
@@ -56,13 +59,34 @@ if ($model->locale && isset(Yii::$app->translations) && class_exists('\wdmg\tran
                     return $output;
                 }
             ],
-            'title:ntext',
-            'description:ntext',
-            'keywords:ntext',
+            [
+                'attribute' => 'title',
+                'format' => 'ntext',
+                'contentOptions' => [
+                    'lang' => ($model->locale ?? Yii::$app->language)
+                ]
+            ],
+            [
+                'attribute' => 'description',
+                'format' => 'ntext',
+                'contentOptions' => [
+                    'lang' => ($model->locale ?? Yii::$app->language)
+                ]
+            ],
+            [
+                'attribute' => 'keywords',
+                'format' => 'ntext',
+                'contentOptions' => [
+                    'lang' => ($model->locale ?? Yii::$app->language)
+                ]
+            ],
             [
                 'attribute' => 'posts',
                 'label' => Yii::t('app/modules/blog', 'Posts'),
                 'format' => 'html',
+                'contentOptions' => [
+                    'lang' => ($model->locale ?? Yii::$app->language)
+                ],
                 'value' => function($data) {
                     if ($posts = $data->posts) {
                         return Html::a(count($posts), ['posts/index', 'tag_id' => $data->id]);