+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+## 1.2.*
+ * Lang attribute for HTML tags and inputs
+ 
 ## 1.2.4 (2020-12-29)
  * RBAC implementation